+++ src/app/api/like/route.ts
@@ -5,28 +5,6 @@ import { handleAPIError } from '../lib/handleAPIError';
 import prisma from '../lib/prisma';
 import { apiRes } from '../types';
 
-//ユーザーがいいねした投稿を返すapi
-export const GET = async (req: Request, res: NextResponse) =>
-  handleAPIError(async () => {
-    dbConnect();
-
-    //clerkId
-    const { userId } = auth();
-
-    //Userテーブルのid(sns上でのID)
-    const { id } = await prisma.user.findUniqueOrThrow({
-      where: { clerkId: userId },
-    });
-
-    const likedPost = await prisma.like.findMany({
-      where: {
-        authorId: id,
-      },
-      include: { post: true, author: true },
-    });
-    return NextResponse.json<apiRes>({ message: 'success', data: likedPost }, { status: 200 });
-  });
-
 export const POST = async (req: Request, res: NextResponse) =>
   handleAPIError(async () => {
     dbConnect();
